refactor: Update loading component for saved notes page with lazy loading for images
