+++ app/loading.jsx
@@ -1,4 +1,4 @@
-import LoadingSkeleton from "@/components/LoadingSkeleton/LoadingSkeleton";
+import LoadingSkeleton from "@/components/Skeleton/LoadingSkeleton";
 
 export default function Loading() {
   return (
