+++ components/Footer/Footer.jsx
@@ -1,4 +1,10 @@
-import { GitHubIcon, InstagramIcon, XIcon } from "@/app/assets/svgs/GeneralIcons";
+import {
+  GitHubIcon,
+  InstagramIcon,
+  XIcon,
+} from "@/app/assets/svgs/GeneralIcons";
+import { cn } from "@/lib/utils";
+import { BASE_URL } from "@/utils/constants";
 import { playfair_font } from "@/utils/fonts";
 import Link from "next/link";
 
@@ -6,8 +12,8 @@ const Footer = () => {
   return (
     <>
       <div className="max-w-3xl mx-auto px-4 flex flex-col gap-1 md:flex-row md:justify-between items-center py-2 md:py-3">
-        <Link href="https://dedxnotes.vercel.app/">
-          <h1 className={`font-bold text-lg md:text-xl ${playfair_font}`}>
+        <Link href={BASE_URL}>
+          <h1 className={cn("font-bold text-lg md:text-xl", playfair_font)}>
             Dedx
             <span className="bg-gradient-to-r from-red-500 to-pink-500 bg-clip-text text-transparent">
               Notes
